requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-login==0.6.2
flask-migrate==4.0.4
flask-cors==3.0.10
werkzeug==2.3.4
python-dotenv==1.0.0
email-validator==2.0.0
stripe==5.4.0
Pillow==9.5.0
gunicorn==20.1.0
psycopg2-binary==2.9.6